requirements.txt
openai==1.16.2
azure-cognitiveservices-speech==1.24.2
requests==2.28.1
python-dotenv==0.19.1
inquirer==3.2.4
playsound==1.2.2
